biomedical-classification-app/requirements.txt
streamlit>=1.24.0
tensorflow>=2.13.0
librosa>=0.10.0
numpy>=1.24.3
pillow>=9.5.0
opencv-python>=4.7.0
ultralytics>=8.0.0
streamlit==1.26.0
tensorflow==2.12.0
librosa==0.10.0
pillow==10.0.0
opencv-python==4.8.0.76
ultralytics==8.0.20
numpy==1.27.0
